+++ crates/sweet/examples/macro.rs
@@ -6,5 +6,5 @@ async fn foo() {}
 sweet! {
 	it{}
 	it{foo().await;}
-	it nonSend{foo().await;}
+	it non_send{foo().await;}
 }