+++ crates/sweet/macros/src/lib.rs
@@ -6,13 +6,15 @@ use syn::parse_macro_input;
 #[proc_macro_attribute]
 pub fn sweet_test(attr: TokenStream, input: TokenStream) -> TokenStream {
 	TestCaseAttr::parse(attr, input)
+		.unwrap_or_else(syn::Error::into_compile_error)
+		.into()
 }
 #[proc_macro]
 pub fn sweet(input: TokenStream) -> TokenStream {
 	parse_macro_input!(input as SuiteFunc).out
 }
 
-#[proc_macro]
-pub fn test(input: TokenStream) -> TokenStream {
-	parse_macro_input!(input as TestCaseFunc).out
-}
+// #[proc_macro]
+// pub fn test(input: TokenStream) -> TokenStream {
+// 	parse_macro_input!(input as TestCaseFunc).out
+// }