+++ crates/sweet/macros/src/macros/build_output_native.rs
@@ -1,10 +1,9 @@
 use super::*;
-use proc_macro2::Group;
 use proc_macro2::TokenStream;
 use quote::quote;
 
 pub fn parse_test_case_native(
-	func: &Group,
+	func: &TokenStream,
 	flags: &TestCaseFlags,
 ) -> TokenStream {
 	let config = flags.to_config();
@@ -23,8 +22,8 @@ pub fn parse_test_case_native(
 	)
 }
 
-fn parse_func_native(func: &Group, flags: &TestCaseFlags) -> TokenStream {
-	let contains_await = _contains_await(func.stream());
+fn parse_func_native(func: &TokenStream, flags: &TestCaseFlags) -> TokenStream {
+	let contains_await = contains_await(func);
 	if contains_await {
 		if flags.non_send {
 			quote!(sweet::TestCaseNativeFunc::Series(||{