+++ crates/gamai/test/common/sequence.rs
@@ -7,8 +7,10 @@ pub fn works() -> Result<()> {
 	let my_tree = || {
 		tree! {
 			<sequence apply_deferred>
-				<node_always_succeed/>
-				<node_always_succeed/>
+				<node_always_succeed apply_deferred/>
+				<node_always_succeed apply_deferred/>
+				<node_always_succeed apply_deferred/>
+				<node_always_succeed apply_deferred/>
 			</sequence>
 		}
 	};
@@ -22,17 +24,22 @@ pub fn works() -> Result<()> {
 
 	let out = my_tree.get_recursive::<Running>(&app.world, entity);
 	expect(out.value).to_be_some()?;
+	let out = my_tree.get_recursive::<NodeState>(&app.world, entity);
 	expect(out.children[0].value).to_be_some()?;
-
+	expect(out.children[1].value).to_be_none()?;
+	
 	app.update();
-	// this looks wrong, is apply_deferred / system ordering broken?
+	
+	let out = my_tree.get_recursive::<Running>(&app.world, entity);
+	expect(out.value).to_be_some()?;
+	let out = my_tree.get_recursive::<NodeState>(&app.world, entity);
+	expect(out.children[1].value).to_be_some()?;
+	
+	app.update();
 	// app.update();
-
-	// let out = my_tree.get_recursive::<Running>(&app.world, entity);
-	// expect(out.value).to_be_some()?;
-	// expect(out.children[0].value).to_be_none()?;
-	// // expect(out.children[0].value).to_be(Some(&NodeState::Success))?;
-	// expect(out.children[1].value).to_be_some()?;
-
+	// app.update();
+	// app.update();
+	let out = my_tree.get_recursive::<Running>(&app.world, entity);
+	expect(out.value).to_be_some()?;
 	Ok(())
 }