+++ docs/src/sweet/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## TODO
+- attribute macros
+- Kill long-running tests
+
 ## `0.1.31` - 13/09/2023
 
 - Headless in-browser tests