+++ docs/src/sweet/changelog.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## TODO
-- attribute macros
 - Kill long-running tests
+
+## `main`
+- attribute macros
 
 ## `0.1.31` - 13/09/2023